feat:#52 댓글 삭제 alert 처리

+++ src/components/detail/MovieComments.tsx
@@ -4,6 +4,8 @@ import WrapperBox from '@/components/detail/WrapperBox';
 import { Button } from '@/components/ui/button';
 import { deleteMovieComment } from '@/services/detail/serviceComments';
 import { useAuthStore } from '@/store/useAuthStore';
+import { ALERT_TYPE } from '@/constants/alertType';
+import { openAlert } from '@/lib/openAlert';
 
 interface Props {
   comments: Comment[];
@@ -12,11 +14,26 @@ interface Props {
 
 const MovieComments = ({ comments, onDelete }: Props) => {
   const { isSignedIn, user } = useAuthStore();
+  const { ERROR, SUCCESS, INFO, WARNING } = ALERT_TYPE;
 
-  const handleDeleteComment = async (e: React.MouseEvent<HTMLButtonElement>) => {
-    const commentId = e.currentTarget.value;
-    await deleteMovieComment(commentId);
-    onDelete(commentId);
+  const handleDeleteComment = async (event: React.MouseEvent<HTMLButtonElement>) => {
+    try {
+      const commentId = event.currentTarget.value;
+      const res = await openAlert({
+        type: WARNING,
+        title: '댓글',
+        text: '댓글을 삭제하시겠습니까?',
+        buttonText: '삭제',
+      });
+
+      if (res!.isConfirmed) {
+        await deleteMovieComment(commentId);
+        openAlert({ type: INFO, title: '댓글', text: '댓글이 삭제됐습니다!' });
+        onDelete(commentId);
+      }
+    } catch (error: Error | any) {
+      openAlert({ type: ERROR, text: error.message });
+    }
   };
 
   return (
